@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <TieredCompilation>true</TieredCompilation>
   </PropertyGroup>
   <!-- Supressing Net Core migration warnings-->
@@ -8,6 +8,9 @@
     <NoWarn>$(NoWarn);NU1608</NoWarn>
     <NoWarn>$(NoWarn);NU1701</NoWarn>
     <NoWarn>$(NoWarn);NU1604</NoWarn>
+    <NoWarn>$(NoWarn);NU1591</NoWarn>
+    <NoWarn>$(NoWarn);NU1702</NoWarn>
+    <NoWarn>$(NoWarn);CS1591</NoWarn>
     <SuppressNETCoreSdkPreviewMessage>true</SuppressNETCoreSdkPreviewMessage>
   </PropertyGroup>
   <PropertyGroup>
@@ -20,13 +23,16 @@
     <Message Importance="High" Text="%0a+++ Building KuduLite +++%0a%0a" />
   </Target>
   <ItemGroup>
+    <PackageReference Include="AspNetCore.RouteAnalyzer" Version="0.5.3" />
+    <PackageReference Include="KubernetesClient" Version="1.5.25" />
     <PackageReference Include="log4net" Version="2.0.8" />
-    <PackageReference Include="Microsoft.AspNetCore.App"/>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Proxy"/>
-    <PackageReference Include="Microsoft.Diagnostics.Runtime"/>
+    <PackageReference Include="Microsoft.AspNetCore.Proxy" />
+    <PackageReference Include="Microsoft.Diagnostics.Runtime" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
     <PackageReference Include="XmlSettings" Version="0.1.3" />
+    <PackageReference Include="System.Runtime.Caching" Version="4.5.0" />
   </ItemGroup>
   <Target Name="_IncludePrePublishGeneratedContent" BeforeTargets="GetCopyToPublishDirectoryItems" Condition=" '$(EnableDefaultItems)' == 'true' And '$(EnableDefaultContentItems)' == 'true' " />
   <ItemGroup>
@@ -37,14 +43,11 @@
     <!-- Odds and ends for the Scripts\ directory -->
     <ContentWithTargetPath Include="$(NuGetPackageRoot)KuduSync.NET\**\tools\*" TargetPath="Scripts\%(FileName)%(Extension)" CopyToOutputDirectory="Always" />
     <ContentWithTargetPath Include="$(NuGetPackageRoot)KuduHandles\**\KuduHandles.exe" TargetPath="Scripts\KuduHandles.exe" CopyToOutputDirectory="Always" />
-    <!-- TODO this relative path may change -->
   </ItemGroup>
   <ItemGroup>
     <Folder Include="wwwroot\Content\Styles\" />
     <Folder Include="wwwroot\Content\Scripts\" />
   </ItemGroup>
-  <!-- CORE TODO this used to have a "BuildOutputsUpdated" condition that used compile timestamps. Not sure if they're still available. -->
-  <!-- CORE TODO I think the best way to make this work right with publish is to restore to the source folder and then add a dynamic ContentWithTargetPath -->
   <!-- Run updateNodeModules to install kuduscript -->  
   <Target Name="PutNodeModulesInBinDir" AfterTargets="Build">    
 	  <Exec Command="$(ProjectDir)\updateNodeModules.cmd &quot;$(PublishDir)&quot;" Condition="'$(OS)' == 'Windows_NT' AND  '$(Configuration)'=='Release'" />
