@@ -5,9 +5,9 @@ namespace Kudu.Core.Deployment
 {
     public interface IDeploymentStatusManager
     {
-        IDeploymentStatusFile Create(string id);
-        IDeploymentStatusFile Open(string id);
-        void Delete(string id);
+        IDeploymentStatusFile Create(string id, IEnvironment environment);
+        IDeploymentStatusFile Open(string id, IEnvironment environment);
+        void Delete(string id, IEnvironment environment);
 
         IOperationLock Lock { get; }
 
