@@ -10,6 +10,8 @@ using Kudu.Core.Infrastructure;
 using Microsoft.Win32;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Extensions;
+using Kudu.Core.Settings;
+using Kudu.Core.K8SE;
 
 namespace Kudu.Core
 {
@@ -39,6 +41,10 @@ namespace Kudu.Core
         private readonly string _jobsBinariesPath;
         private readonly string _sitePackagesPath;
         private readonly string _secondaryJobsBinariesPath;
+        private readonly string _k8seAppName;
+        private readonly string _k8seAppNamespace;
+        private readonly string _k8seAppType;
+
 
         // This ctor is used only in unit tests
         public Environment(
@@ -58,7 +64,10 @@ namespace Kudu.Core
                 string siteExtensionSettingsPath,
                 string sitePackagesPath,
                 string requestId,
-                IHttpContextAccessor httpContextAccessor)
+                IHttpContextAccessor httpContextAccessor,
+                string k8seAppName = null,
+                string k8seAppNamespace = null,
+                string k8seAppType = null)
         {
             if (repositoryPath == null)
             {
@@ -77,8 +86,8 @@ namespace Kudu.Core
             _diagnosticsPath = diagnosticsPath;
             _locksPath = locksPath;
             _sshKeyPath = sshKeyPath;
-            Console.WriteLine("Root Path : "+rootPath);
-            Console.WriteLine("SSH Key Path : "+_sshKeyPath);
+            Console.WriteLine("Root Path : " + rootPath);
+            Console.WriteLine("SSH Key Path : " + _sshKeyPath);
             _scriptPath = scriptPath;
             _nodeModulesPath = nodeModulesPath;
 
@@ -97,6 +106,9 @@ namespace Kudu.Core
 
             RequestId = !string.IsNullOrEmpty(requestId) ? requestId : Guid.Empty.ToString();
             _httpContextAccessor = httpContextAccessor;
+            _k8seAppName = k8seAppName;
+            _k8seAppNamespace = k8seAppNamespace;
+            _k8seAppType = k8seAppType;
         }
 
         public Environment(
@@ -105,7 +117,10 @@ namespace Kudu.Core
                 string repositoryPath,
                 string requestId,
                 string kuduConsoleFullPath,
-                IHttpContextAccessor httpContextAccessor)
+                IHttpContextAccessor httpContextAccessor,
+                string k8seAppName = null,
+                string k8seAppNamespace = null,
+                string k8seAppType = null)
         {
             RootPath = rootPath;
 
@@ -120,7 +135,10 @@ namespace Kudu.Core
             _siteExtensionSettingsPath = Path.Combine(SiteRootPath, Constants.SiteExtensionsCachePath);
             _diagnosticsPath = Path.Combine(SiteRootPath, Constants.DiagnosticsPath);
             _locksPath = Path.Combine(SiteRootPath, Constants.LocksPath);
-            
+            _k8seAppName = k8seAppName;
+            _k8seAppNamespace = k8seAppNamespace;
+            _k8seAppType = k8seAppType;
+
             if (OSDetector.IsOnWindows())
             {
                 _sshKeyPath = Path.Combine(rootPath, Constants.SSHKeyPath);
@@ -128,7 +146,12 @@ namespace Kudu.Core
             else
             {
                 // in linux, rootPath is "/home", while .ssh folder need to under "/home/{user}"
-                _sshKeyPath = Path.Combine(rootPath, System.Environment.GetEnvironmentVariable("KUDU_RUN_USER"), Constants.SSHKeyPath);
+                string path2 = System.Environment.GetEnvironmentVariable("KUDU_RUN_USER");
+                if (path2 == null || path2.Equals(""))
+                {
+                    path2 = "root";
+                }
+                _sshKeyPath = Path.Combine(rootPath, path2, Constants.SSHKeyPath);
             }
             _scriptPath = Path.Combine(binPath, Constants.ScriptsPath);
             _nodeModulesPath = Path.Combine(binPath, Constants.NodeModulesPath);
@@ -152,7 +175,7 @@ namespace Kudu.Core
                 _jobsBinariesPath = Path.Combine(_webRootPath, userDefinedWebJobRoot);
             }
             _sitePackagesPath = Path.Combine(_dataPath, Constants.SitePackages);
-            
+
             RequestId = !string.IsNullOrEmpty(requestId) ? requestId : Guid.Empty.ToString();
 
             _httpContextAccessor = httpContextAccessor;
@@ -224,7 +247,7 @@ namespace Kudu.Core
         public string RootPath
         {
             get;
-            private set;
+            set;
         }
 
         public string SiteRootPath
@@ -343,7 +366,7 @@ namespace Kudu.Core
                 return this.WebRootPath;
             }
         }
-        
+
         public string SitePackagesPath
         {
             get
@@ -385,11 +408,21 @@ namespace Kudu.Core
             private set;
         }
 
+        public bool IsOnLinuxConsumption
+        {
+            get
+            {
+                bool isOnAppService = !string.IsNullOrEmpty(System.Environment.GetEnvironmentVariable(Constants.AzureWebsiteInstanceId));
+                bool isOnLinuxContainer = !string.IsNullOrEmpty(System.Environment.GetEnvironmentVariable(Constants.ContainerName));
+                return isOnLinuxContainer && !isOnAppService;
+            }
+        }
+
         public string KuduConsoleFullPath { get; }
 
         public static bool IsAzureEnvironment()
         {
-            return !String.IsNullOrEmpty(System.Environment.GetEnvironmentVariable("WEBSITE_INSTANCE_ID"));
+            return !String.IsNullOrEmpty(System.Environment.GetEnvironmentVariable(Constants.AzureWebsiteInstanceId));
         }
 
         public static bool SkipSslValidation
@@ -417,6 +450,65 @@ namespace Kudu.Core
                 }
 
                 return skipSslValidation == "1";
+            }
+        }
+
+
+        public static string ContainerName
+        {
+            get
+            {
+                return System.Environment.GetEnvironmentVariable(Constants.ContainerName)?.ToLowerInvariant();
+            }
+        }
+
+        public static string StampName
+        {
+            get
+            {
+                return System.Environment.GetEnvironmentVariable(Constants.WebSiteHomeStampName)?.ToLowerInvariant();
+            }
+        }
+
+        public static string TenantId
+        {
+            get
+            {
+                return System.Environment.GetEnvironmentVariable(Constants.WebSiteStampDeploymentId)?.ToLowerInvariant();
+            }
+        }
+
+        public bool IsK8SEApp
+        {
+            get
+            {
+                return K8SEDeploymentHelper.IsK8SEEnvironment();
+            }
+        }
+
+        public string CurrId { get; set; }
+
+        public string K8SEAppName
+        {
+            get
+            {
+                return _k8seAppName;
+            }
+        }
+
+        public string K8SEAppNamespace
+        {
+            get
+            {
+                return _k8seAppNamespace;
+            }
+        }
+
+        public string K8SEAppType
+        {
+            get
+            {
+                return _k8seAppType;
             }
         }
 
