@@ -1,11 +1,13 @@
 using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
- namespace Kudu.Core.Helpers
+using Kudu.Contracts.Settings;
+using Microsoft.AspNetCore.Authentication;
+
+namespace Kudu.Core.Helpers
 {
     public static class SimpleWebTokenHelper
     {
@@ -15,19 +17,25 @@ using System.Text;
         /// The SWT is then returned as an encrypted string
         /// </summary>
         /// <param name="validUntil">Datetime for when the token should expire</param>
+        /// <param name="key">Optional key to encrypt the token with</param>
         /// <returns>a SWT signed by this app</returns>
-        public static string CreateToken(DateTime validUntil) => Encrypt($"exp={validUntil.Ticks}");
+        public static string CreateToken(DateTime validUntil, byte[] key = null) => Encrypt($"exp={validUntil.Ticks}", key);
          [SuppressMessage("Microsoft.Usage", "CA2202:Object 'cipherStream' and 'cryptoStream' can be disposed mo re than once",
             Justification = "MemoeryStream, CryptoStream, and BinaryWriter handle multiple disposal correctly. The alternative is pretty ugly code for clearing each variable, checking for null, and manual dispose.")]
-        private static string Encrypt(string value)
+        public static string Encrypt(string value, byte[] key = null)
         {
-            using (var aes = new AesManaged { Key = GetWebSiteAuthEncryptionKey() })
+            if (key == null)
+            {
+                TryGetEncryptionKey(SettingsKeys.AuthEncryptionKey, out key);
+            }
+
+            using (var aes = new AesManaged { Key = key })
             {
                 // IV is always generated for the key every time
                 aes.GenerateIV();
                 var input = Encoding.UTF8.GetBytes(value);
                 var iv = Convert.ToBase64String(aes.IV);
-                 using (var encrypter = aes.CreateEncryptor(aes.Key, aes.IV))
+                using (var encrypter = aes.CreateEncryptor(aes.Key, aes.IV))
                 using (var cipherStream = new MemoryStream())
                 {
                     using (var cryptoStream = new CryptoStream(cipherStream, encrypter, CryptoStreamMode.Write))
@@ -41,21 +49,114 @@ using System.Text;
                 }
             }
         }
-         private static string GetSHA256Base64String(byte[] key)
+
+        public static bool TryValidateToken(string token, ISystemClock systemClock)
+        {
+            try
+            {
+                return ValidateToken(token, systemClock);
+            }
+            catch
+            {
+                return false;
+            }
+        }
+
+        public static bool ValidateToken(string token, ISystemClock systemClock)
+        {
+            // Use WebSiteAuthEncryptionKey if available else fallback to ContainerEncryptionKey.
+            // Until the container is specialized to a specific site WebSiteAuthEncryptionKey will not be available.
+            byte[] key;
+            if (!TryGetEncryptionKey(SettingsKeys.AuthEncryptionKey, out key, false))
+            {
+                TryGetEncryptionKey(SettingsKeys.ContainerEncryptionKey, out key);
+            }
+
+            var data = Decrypt(key, token);
+
+            var parsedToken = data
+                // token = key1=value1;key2=value2
+                .Split(';', StringSplitOptions.RemoveEmptyEntries)
+                 // ["key1=value1", "key2=value2"]
+                 .Select(v => v.Split('=', StringSplitOptions.RemoveEmptyEntries))
+                 // [["key1", "value1"], ["key2", "value2"]]
+                 .ToDictionary(k => k[0], v => v[1]);
+
+            return parsedToken.ContainsKey("exp") && systemClock.UtcNow.UtcDateTime < new DateTime(long.Parse(parsedToken["exp"]));
+        }
+
+        public static string Decrypt(byte[] encryptionKey, string value)
+        {
+            var parts = value.Split(new[] { '.' }, StringSplitOptions.RemoveEmptyEntries);
+            if (parts.Length != 2 && parts.Length != 3)
+            {
+                throw new InvalidOperationException("Malformed token.");
+            }
+
+            var iv = Convert.FromBase64String(parts[0]);
+            var data = Convert.FromBase64String(parts[1]);
+            var base64KeyHash = parts.Length == 3 ? parts[2] : null;
+
+            if (!string.IsNullOrEmpty(base64KeyHash) && !string.Equals(GetSHA256Base64String(encryptionKey), base64KeyHash))
+            {
+                throw new InvalidOperationException(string.Format("Key with hash {0} does not exist.", base64KeyHash));
+            }
+
+            using (var aes = new AesManaged { Key = encryptionKey })
+            {
+                using (var ms = new MemoryStream())
+                {
+                    using (var cs = new CryptoStream(ms, aes.CreateDecryptor(aes.Key, iv), CryptoStreamMode.Write))
+                    using (var binaryWriter = new BinaryWriter(cs))
+                    {
+                        binaryWriter.Write(data, 0, data.Length);
+                    }
+
+                    return Encoding.UTF8.GetString(ms.ToArray());
+                }
+            }
+        }
+
+        private static bool TryGetEncryptionKey(string keyName, out byte[] encryptionKey, bool throwIfFailed = true)
+        {
+            encryptionKey = null;
+            var hexOrBase64 = System.Environment.GetEnvironmentVariable(keyName);
+            if (string.IsNullOrEmpty(hexOrBase64))
+            {
+                if (throwIfFailed)
+                {
+                    throw new InvalidOperationException($"No {keyName} defined in the environment");
+                }
+
+                return false;
+            }
+
+            encryptionKey = hexOrBase64.ToKeyBytes();
+
+            return true;
+        }
+
+        private static string GetSHA256Base64String(byte[] key)
         {
             using (var sha256 = new SHA256Managed())
             {
                 return Convert.ToBase64String(sha256.ComputeHash(key));
             }
         }
-         private static byte[] GetWebSiteAuthEncryptionKey()
+
+        private static byte[] GetWebSiteAuthEncryptionKey()
         {
             var hexOrBase64 = System.Environment.GetEnvironmentVariable(Constants.SiteAuthEncryptionKey);
             if (string.IsNullOrEmpty(hexOrBase64))
             {
                 throw new InvalidOperationException($"No {Constants.SiteAuthEncryptionKey} defined in the environment");
             }
-            
+
+            return hexOrBase64.ToKeyBytes();
+        }
+
+        public static byte[] ToKeyBytes(this string hexOrBase64)
+        {
             // only support 32 bytes (256 bits) key length
             if (hexOrBase64.Length == 64)
             {
@@ -64,6 +165,7 @@ using System.Text;
                     .Select(x => Convert.ToByte(hexOrBase64.Substring(x, 2), 16))
                     .ToArray();
             }
+
             return Convert.FromBase64String(hexOrBase64);
         }
     }
