@@ -1,12 +1,11 @@
 ﻿using System;
-using Microsoft.AspNetCore.Routing;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Routing;
 
 namespace Kudu.Services.Web.Infrastructure
 {
-
     // CORE NOTE This was renamed/reworked from RouteCollectionExtensions
-    
+
     /// <summary>
     /// 
     /// </summary>
@@ -16,34 +15,42 @@ namespace Kudu.Services.Web.Infrastructure
     {
         public const string DeprecatedKey = "deprecated";
 
-        public static void MapHttpWebJobsRoute(this IRouteBuilder routes, string name, string jobType, string routeTemplate, object defaults, object constraints = null)
+        public static void MapHttpWebJobsRoute(this IRouteBuilder routes, string name, string jobType,
+            string routeTemplate, object defaults, object constraints = null)
         {
             // e.g. api/continuouswebjobs/foo
-            routes.MapHttpRoute(name, String.Format("api/{0}webjobs{1}", jobType, routeTemplate), defaults, constraints, deprecated: false);
+            routes.MapHttpRoute(name, String.Format("api/{0}webjobs{1}", jobType, routeTemplate), defaults, constraints,
+                deprecated: false);
 
             // e.g. api/triggeredjobs/foo/history/17
-            routes.MapHttpRoute(name + "-dep", String.Format("api/{0}jobs{1}", jobType, routeTemplate), defaults, constraints, deprecated: true);
+            routes.MapHttpRoute(name + "-dep", String.Format("api/{0}jobs{1}", jobType, routeTemplate), defaults,
+                constraints, deprecated: true);
 
             // e.g. jobs/triggered/foo/history/17 and api/jobs/triggered/foo/history/17
-            routes.MapHttpRouteDual(name + "-old", String.Format("jobs/{0}{1}", jobType, routeTemplate), defaults, constraints, bothDeprecated: true);
+            routes.MapHttpRouteDual(name + "-old", String.Format("jobs/{0}{1}", jobType, routeTemplate), defaults,
+                constraints, bothDeprecated: true);
         }
 
-        public static void MapHttpProcessesRoute(this IRouteBuilder routes, string name, string routeTemplate, object defaults, object constraints = null)
+        public static void MapHttpProcessesRoute(this IRouteBuilder routes, string name, string routeTemplate,
+            object defaults, object constraints = null)
         {
             // e.g. api/processes/3958/dump
-            routes.MapHttpRoute(name + "-direct", "api/processes" + routeTemplate, defaults, constraints, deprecated: false);
+            routes.MapHttpRoute(name + "-direct", "api/processes" + routeTemplate, defaults, constraints,
+                deprecated: false);
 
             // e.g. api/diagnostics/processes/4845
             routes.MapHttpRouteDual(name, "diagnostics/processes" + routeTemplate, defaults, constraints);
         }
 
-        public static void MapHttpRouteDual(this IRouteBuilder routes, string name, string routeTemplate, object defaults, object constraints = null, bool bothDeprecated = false)
+        public static void MapHttpRouteDual(this IRouteBuilder routes, string name, string routeTemplate,
+            object defaults, object constraints = null, bool bothDeprecated = false)
         {
             routes.MapHttpRoute(name + "-dep", routeTemplate, defaults, constraints, deprecated: true);
             routes.MapHttpRoute(name, "api/" + routeTemplate, defaults, constraints, deprecated: bothDeprecated);
         }
 
-        public static void MapHttpRoute(this IRouteBuilder routes, string name, string routeTemplate, object defaults, object constraints, bool deprecated)
+        public static void MapHttpRoute(this IRouteBuilder routes, string name, string routeTemplate, object defaults,
+            object constraints, bool deprecated)
         {
             // CORE TODO Note that the only place that the deprecated datatoken is used, it only checks to see
             // if the key is *there*, not the bool value of it, so this is a little awkward looking due to the way
@@ -51,7 +58,7 @@ namespace Kudu.Services.Web.Infrastructure
             if (deprecated)
             {
                 name += "-dep";
-                routes.MapRoute(name, routeTemplate, defaults, constraints, new { DeprecatedKey = true });
+                routes.MapRoute(name, routeTemplate, defaults, constraints, new {DeprecatedKey = true});
             }
             else
             {
