@@ -63,7 +63,7 @@
             </li>
         }
         
-        @foreach (KeyValuePair<string, string> kv in _settingsManager.GetValues())
+        @foreach (KeyValuePair<string, string> kv in _settingsManager.GetValues((IDictionary<string, string>)HttpContext.Items["appSettings"]))
         {
             
                 if (kv.Value != null)
