@@ -30,7 +30,6 @@
         Assembly assembly;
         try
         {
-            //return "test";
             assembly = Assembly.Load("Microsoft.Web.Hosting, Version=7.1.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35");
             var fileVersionInfo = System.Diagnostics.FileVersionInfo.GetVersionInfo(assembly.Location);
             return fileVersionInfo.ProductVersion;
@@ -48,10 +47,10 @@
 </style>
 <div class="container">
     @{
-        //string commitFile = Server.MapPath("~/commit.txt");
         string commitFile = System.IO.Path.Combine(hostingEnvironment.WebRootPath, "commit.txt");
         string sha = System.IO.File.Exists(commitFile) ? System.IO.File.ReadAllText(commitFile).Trim() : null;
-        var version = typeof(Kudu.Services.Web.Tracing.TraceMiddleware).Assembly.GetName().Version;
+        //var version = typeof(Kudu.Services.Web.Tracing.TraceMiddleware).Assembly.GetName().Version;
+        var version = Constants.KuduBuild;
     }
     <h3>Environment</h3>
     <div class="row">
@@ -69,24 +68,12 @@
                 }
         </div>
     </div>
-    @if (appServiceVersion != null)
-    {
-        <div class="row">
-            <div class="col-xs-2">
-                <strong>Azure App Service</strong>
-            </div>
-            <div>
-                @GetAppServiceVersion()
-            </div>
-        </div>
-    }
     <div class="row">
         <div class="col-xs-2">
             <strong>Site up time</strong>
         </div>
         <div>
             @Kudu.Services.Web.Tracing.TraceMiddleware.UpTime.ToString(@"dd\.hh\:mm\:ss")
-            12345
         </div>
     </div>
     <div class="row">
@@ -118,9 +105,6 @@
         </li>
         <li>
             <a href="api/vfs">Files</a>
-        </li>
-        <li>
-            <a href="api/logs/docker">Current Docker logs</a> (<a href="api/logs/docker/zip">Download as zip</a>)
         </li>
     </ul>
     <h3>More information about Kudu can be found on the <a href="https://github.com/projectkudu/kudu/wiki">wiki</a>.</h3>
