@@ -62,8 +62,8 @@
                 @name = @System.Configuration.ConfigurationManager.AppSettings[name]
             </li>
         }
-        
-        @foreach (KeyValuePair<string, string> kv in _settingsManager.GetValues())
+
+        @foreach (KeyValuePair<string, string> kv in _settingsManager.GetValues((IDictionary<string, string>)HttpContext.Items["appSettings"]))
         {
             
                 if (kv.Value != null)
