@@ -1,9 +1,12 @@
 
 @page
 @using Kudu.Core.Deployment
+
+@using Kudu.Core
 @using Microsoft.AspNetCore.Hosting
 @inject IHostingEnvironment hostingEnvironment
 @inject IDeploymentManager deploymentManager
+@inject Microsoft.AspNetCore.Http.IHttpContextAccessor accessor
 
 @{
     ViewBag.Title = "title";
@@ -11,7 +14,9 @@
 }
 
 <style>
-    .bg-grad { background-image: linear-gradient(to right, #06B0BA, #2F3B61); }
+    .bg-grad {
+        background-image: linear-gradient(to right, #06B0BA, #2F3B61);
+    }
 </style>
 
 <script type="text/javascript">
@@ -33,10 +38,7 @@
 </script>
 
 @{
-    //string commitFile = Server.MapPath("~/commit.txt");
-    var commitFile = System.IO.Path.Combine(hostingEnvironment.WebRootPath, "commit.txt");
-    var sha = System.IO.File.Exists(commitFile) ? System.IO.File.ReadAllText(commitFile).Trim() : null;
-    var workerIdHash = System.Environment.GetEnvironmentVariable("WEBSITE_INSTANCE_ID");
+    var workerIdHash = System.Environment.GetEnvironmentVariable(Constants.AzureWebsiteInstanceId);
     var workerIdShortHash = workerIdHash?.Substring(0, 12);
     const string version = Constants.KuduBuild;
     var showLearningBanner = !deploymentManager.GetResults().Any();
@@ -48,19 +50,31 @@
     <div class="row border-between">
         <div class="col-md-7">
             <div class="row align-middle" style="margin-left: 1px">
-                <h3>Kudu<span style="font-weight: 200">Lite</span></h3>@version</div>
-            <small>Site UpTime: <i class="fas fa-circle" style="color: #26C281; font-size: x-small; margin-right: 5px"></i>@Tracing.TraceMiddleware.UpTime.ToString(@"dd\.hh\:mm\:ss") | Site Folder: @PathResolver.ResolveRootPath() | Temp Folder: @System.IO.Path.GetTempPath()</small>
-            <br> <small onclick="clickLink(this)" id="curr-instance" title="@workerIdHash">Kudu Attached to Instance: &nbsp;<i class="fas fa-desktop fa-style"></i>&nbsp;<a href="#">@workerIdShortHash </a>| <i class="fas fa-network-wired fa-style"></i> <a href="#"> Switch Instance</a></small>
+                <h3>Kudu<span style="font-weight: 200">Lite</span></h3>@version
+            </div>
+            <small>Site UpTime: <i class="fas fa-circle" style="color: #26C281; font-size: x-small; margin-right: 5px"></i>@Tracing.TraceMiddleware.UpTime.ToString(@"dd\.hh\:mm\:ss") | Site Folder: @(((IEnvironment)accessor.HttpContext.Items["environment"]).SiteRootPath) | Temp Folder: @System.IO.Path.GetTempPath()</small>
+            <br> <small onclick="clickLink(this)" id="curr-instance" title="@workerIdHash">
+                Kudu Attached to Instance: &nbsp;<i class="fas fa-desktop fa-style"></i>&nbsp;<a href="#">@workerIdShortHash </a>| <span class="dropdown">
+                    <button class="btn btn-sm dropdown-toggle" type="button" style="background-color: rgb(244, 245, 247);border: none" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                        <i class="fas fa-network-wired fa-style text-sm"></i> <a href="#"> Switch Instance</a>
+                    </button>
+                    <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
+                        <a class="dropdown-item" href="#">Instance 1</a>
+                        <a class="dropdown-item" href="#">Instance 2</a>
+                        <a class="dropdown-item" href="#">Instance 3</a>
+                    </div>
+                </span>
+            </small>
         </div>
         <div class="col-md-5" id="is-deploying-banner" style="display: none">
             <div>
                 <div class="col-md-10 row-fluid">
-                    <h6 style="font-weight: 300"> Deployment Underway</h6>
+                    <h6 id="lock_msg" style="font-weight: 300"></h6>
                     <h6 class="badge badge-secondary inconsolata align-left" style="background-color: #343a40; color: #fafafa" id="deployment-status">...</h6>
                 </div>
                 <div class="col-md-2 row-fluid">
                     <div class="centering text-center">
-                        <a class="btn btn-primary btn-sm align-middle" href="#" role="button">Stream Logs</a>
+                        <a class="btn btn-primary btn-sm align-middle" href="/api/logstream" role="button">Stream Logs</a>
                     </div>
                 </div>
             </div>
@@ -69,13 +83,20 @@
             </div>
         </div>
         <div class="col-md-5" id="just-deployed-banner" style="display: none">
-            <div class="col-md-9 row-fluid">
-                Deployment Successful
-            </div>
-            <div class="col-md-3 row-fluid">
-                <div class="centering text-center">
-                    <h2 style="font-weight: 300"> <i class="fas fa-check-circle" style="color: #26c281"></i></h2>
+
+            <div class="col-md-10 row-fluid">
+                <h6 style="font-weight: 400; margin-bottom: -10px">
+                    Deployment &nbsp;
+                    <p class="badge badge-secondary inconsolata align-left" style="background-color: #343a40; color: #fafafa"> Successful <i class="fas fa-check-circle" style="color: #26C281;"></i>&nbsp;</p>
+                </h6>
+                <div class="row-fluid">
+                    <small style="color: #42526e"> Performing an Overlapped Recycle of your WebApp. Your changes would reflect in <span id="statusStrTimer"></span> seconds</small>
                 </div>
+            </div>
+            <div class="col-md-2 row-fluid" style="margin: 0px">
+                <button type="button" class="btn btn-primary btn-sm">
+                    <i class="fas fa-recycle"></i> <span id="timer" class="badge badge-light"></span> Seconds
+                </button>
             </div>
         </div>
         <div class="col-md-5" id="last-deployed-banner" style="@lastDeployedDisplayVal">
@@ -97,21 +118,12 @@
         </div>
         <div class="col-md-5" id="appsvc-welcome-banner" style="@welcomeBannerDisplayVal">
             <div class="container" style="height: 100%; width: 100%; display: flex; align-items: center; vertical-align: middle">
-                <div class="row text-center">
-                    <div class="col-sm-12 col-md-4 col-md-offset-0">
-                        <div class="wrapper">
-                            <a class="btn btn-primary btn-sm" href="#" role="button"><i class="fas fa-cloud-upload-alt"></i>Deploy your first app</a>
-                        </div>
+                <div class="row col-md-12 col-sm-12">
+                    <div class="col-lg-6 col-md-12 col-sm-12 mb-10 mt-10 text-center">
+                        <button type="button" onclick="location.href = 'https://go.microsoft.com/fwlink/?linkid=2084231';" class="btn btn-primary btn-sm"><i class="fas fa-cloud-upload-alt"></i>Deploy your first app</button>
                     </div>
-                    <div class="col-sm-12 col-md-4 col-md-offset-0">
-                        <div class="wrapper">
-                            <a class="btn btn-primary btn-sm align-middle bkg-light-gray border-0" href="#" role="button" style="margin-left: -10px"><i class="fab fa-wikipedia-w"></i> Read Kudu Wiki</a>
-                        </div>
-                    </div>
-                    <div class="col-sm-12 col-md-4 col-md-offset-0">
-                        <div class="wrapper">
-                            <a class="btn btn-primary btn-sm align-middle bkg-light-gray border-0" href="#" role="button" style="margin-left: -10px"> Learn Node <i class="fab fa-node-js"></i> Development</a>
-                        </div>
+                    <div class="col-lg-6 col-sm-12 col-md-12 mb-10 mt-10 d-xs-none d-sm-none d-md-block text-center">
+                        <button type="button" onclick="location.href = 'https://docs.microsoft.com/en-us/visualstudio/ide/quickstart-nodejs?view=vs-2019';" class="btn bkg-light-gray border-0 btn-sm"><i class="fas fa-cloud-upload-alt"></i> Learn Node <i class="fab fa-node-js"></i> Development</button>
                     </div>
                 </div>
             </div>
@@ -130,9 +142,6 @@
                         <ul class="list-unstyled" style="text-align: left; line-height: 25px">
                             <li>
                                 <a href="/newui/jsonviewer?view_url=/api/settings">App Settings</a>
-                            </li>
-                            <li>
-                                <a href="/newui/wwwroot">Browse WWWRoot</a>
                             </li>
                             <li>
                                 <a href="/newui/jsonviewer?view_url=/diagnostics/runtime">System Configuration</a>
@@ -164,28 +173,6 @@
                 </div>
             </div>
         </div>
-        <div class="card offset-md-1 col-xs-6 shadow-sm border-0" style="background-color: rgb(244, 245, 247); width: 20.8333%">
-            <div class="container">
-                <div class="row" style="margin: 10px; border-bottom: 1px solid #e3e3e3">
-                    <h6 class="font-weight-normal"><i class="fab fa-docker"></i>&nbsp;Web App Container</h6>
-                </div>
-                <div class="row" style="margin: 10px">
-                    <small>
-                        <ul class="list-unstyled" style="text-align: left; line-height: 25px">
-                            <li>
-                                <a href="/newui/jsonviewer?view_url=/api/logs/docker">Current Container Logs</a>
-                            </li>
-                            <li>
-                                <a href="/api/logs/docker/zip">Download All Container Logs</a>
-                            </li>
-                            <li>
-                                <a href="#">Process Dump</a>
-                            </li>
-                        </ul>
-                    </small>
-                </div>
-            </div>
-        </div>
         <div class="col-md-1" style="width: 8.3333%"></div>
     </div>
 </div>
@@ -196,9 +183,11 @@
         $.ajax({
             url: "/api/isdeploying?trace=false",
             type: "GET",
-            success: function(data) {
+            success: function (data) {
                 console.log("deployment going on" + data["value"]);
-                if (data["value"]) {
+
+                $('#lock_msg').text(data["msg"]);
+                if (data["value"] == 'True') {
                     $("#appsvc-welcome-banner").css("display", "none");
                     $("#last-deployed-banner").css("display", "none");
                     $("#just-deployed-banner").css("display", "none");
@@ -206,7 +195,7 @@
                     $.ajax({
                         url: "/api/deployments/latest?trace=false",
                         type: "GET",
-                        success: function(data) {
+                        success: function (data) {
                             console.log("latest deployment data");
                             console.log(data);
                             wasDeploying = true;
@@ -217,11 +206,9 @@
                     });
                 } else {
                     $('#is-deploying-banner').css("display", "none");
-                    console.log("Trying to show just deployed banner : wasDeploying value is "+wasDeploying);
+                    console.log("Trying to show just deployed banner : wasDeploying value is " + wasDeploying);
                     if (wasDeploying) {
-                        console.log("Before updating UI "+wasDeploying);
                         UpdateLatestDeploymentStatus();
-                        console.log("After updating UI "+wasDeploying);
                     }
                 }
             },
@@ -232,31 +219,30 @@
     }
 
     function UpdateLatestDeploymentStatus() {
-        console.log("Making latest API call "+wasDeploying);
+        console.log("Making latest API call " + wasDeploying);
         $.ajax({
             url: "/api/deployments/latest?trace=false",
             type: "GET",
-            success: function(data) {
+            success: function (data) {
                 if (!data["status"]) {
                     return;
                 }
-                console.log("API call done "+wasDeploying);
+                console.log("API call done " + wasDeploying);
                 $("#appsvc-welcome-banner").css("display", "none");
-
+                $("#just-deployed-banner").css("display", "none");
+                $("#diagnose-btn").css("display", "none");
                 if (!wasDeploying) {
                     $("#last-deployed-banner").css("display", "block");
                     if (data["status"] === 4) {
                         $("#latest-deployment-id")
                             .html("Successful &nbsp;<i class=\"fas fa-check-circle\" style=\"color: #26C281;\"></i>")
                             .prop('title', 'Deployment Status');
-                        $("#investigate-deployment")
-                            .html("<i class=\"fab fa-searchengin\"></i> Analyze on AppLens");
                     } else if (data["status"] === 1 || data["status"] === 2) {
                         $("#latest-deployment-id")
-                            .html("Pending &nbsp;<i class=\"fas fa-exclamation-circle\" style=\"color: #ff5e57;\"></i>")
+                            .html("Pending &nbsp;<i class=\"fas fa-exclamation-circle\" style=\"color: #fdcb6e;\"></i>")
                             .prop('title', 'Deployment Status');
                     } else {
-                        $("#diagnose-btn").css("display","inline-block");
+                        $("#diagnose-btn").css("display", "inline-block");
                         $("#latest-deployment-id")
                             .html("Failed &nbsp;<i class=\"fas fa-exclamation-circle\" style=\"color: #ff5e57;\"></i>")
                             .prop('title', 'Deployment Status');
@@ -280,13 +266,28 @@
 
                     $("#latest-deployment-deployer").prop("title", "Deployed by " + data["author"]);
                 } else {
-                    
                     $("#just-deployed-banner").css("display", "inline-block");
+                    setTimeout(timer, 1000);
                     wasDeploying = false;
                 }
             }
         });
     };
+
+    var count = 10;
+    function timer() {
+        count = count - 1;
+        if (count <= 0) {
+            count = 10;
+            document.getElementById("timer").innerHTML = "0";
+            document.getElementById("statusStrTimer").innerHTML = "0";
+            setTimeout(UpdateLatestDeploymentStatus, 1000);
+            return;
+        }
+        document.getElementById("timer").innerHTML = count;
+        document.getElementById("statusStrTimer").innerHTML = count;
+        setTimeout(timer, 1000);
+    }
 
     UpdateLatestDeploymentStatus();
 
@@ -318,8 +319,8 @@
         return Math.floor(seconds) + " seconds";
     }
 
-    setInterval(function() {
-            checkDeployingStatus(); // this will run after every 10 seconds
-        },
+    setInterval(function () {
+        checkDeployingStatus(); // this will run after every 10 seconds
+    },
         10000);
 </script>
