@@ -14,13 +14,14 @@
     {
         <link rel="shortcut icon" type="image/x-icon" href="/linux_favicon.ico" />
         <link rel="icon" type="image/x-icon" href="/linux_favicon.ico" />
-	
+  
     }
 
 <html lang="en">
 
 <head>
-<meta charset="utf-8">
+<meta charset="utf-8"> 
+
 <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
 <meta name="description" content="">
 <meta name="author" content="">
@@ -44,7 +45,22 @@
   <a class="navbar-brand col-sm-3 col-md-2 mr-0" href="/newui"><img src="/Content/Images/AppService.png" height="30" width="30" style="display: inline-block; margin-right: 10px"/>Azure App Service</a>
   <ul class="navbar-nav px-3">
     <li class="nav-item text-nowrap">
-      <a class="nav-link" href="#">Sign out</a>
+                      @{
+                    var email = httpContextAccessor.HttpContext.Request.Headers["X-MS-CLIENT-PRINCIPAL-NAME"].ToString();
+                    var displayName = httpContextAccessor.HttpContext.Request.Headers["X-MS-CLIENT-DISPLAY-NAME"].ToString() ?? email;
+                    if (email != null && email.Contains("@"))
+                    {
+                        <ul class="nav navbar-nav navbar-right" style="position: fixed; right: 65px;">
+                            <li class="dropdown">
+                                <i class="fas fa-user-circle" style="padding:1px;color:#e5e8ec"></i>
+                                <a href="#" data-toggle="dropdown" style="display: inline">@email</a>
+                                <ul class="dropdown-menu">
+                                    <li><a href="/logout">Logout</a></li>
+                                </ul>
+                            </li>
+                        </ul>
+                    }
+                }
     </li>
   </ul>
 </nav>
@@ -70,29 +86,14 @@
               <i class="feather fas fa-terminal clr-icon-navbar"></i> WebSSH
             </a>
           </li>
-          <li class="nav-item">
-            <a class="nav-link" href="/newui/bash">
-              <i class="feather fas fa-laptop-code clr-icon-navbar"></i> Bash
-            </a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="#" title="If no log events are being generated the page may not load.">
-              <i class="feather fas fa-stream clr-icon-navbar"></i> Log stream
-            </a>
-          </li>
-          <li class="nav-item">
+          <li class="nav-item" hidden>
             <a class="nav-link" href="#">
               <i class="feather fas fa-user-md clr-icon-navbar"></i> App Lens
             </a>
           </li>
-          <li class="nav-item">
+          <li class="nav-item" hidden>
             <a class="nav-link" href="#">
               <i class="feather fas fa-globe-americas clr-icon-navbar"></i> Kudu REST APIs
-            </a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="#">
-              <i class="feather fas fa-bolt clr-icon-navbar"></i> VSCode Live Debugging
             </a>
           </li>
         </ul>
@@ -144,8 +145,9 @@
 ================================================== -->
 <!-- Placed at the end of the document so the pages load faster -->
 <script>window.jQuery || document.write('<script src="/assets/js/vendor/jquery-slim.min.js"><\/script>')</script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
+  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
 <!-- Icons -->
 <script src="https://unpkg.com/feather-icons/dist/feather.min.js"></script>
 <script>
