@@ -1,2 +1,3 @@
 @page
-<iframe src="/webssh/host" frameborder="0" scrolling="no" seamless="seamless" style="display:block; width:100%; height:100vh;"></iframe>
+<button onclick="location.href='/instances/any/webssh/host'" style="position: fixed;top:70px; right:30px;" type="button" class="btn btn-outline-light border-0"><i class="fas fa-expand-arrows-alt"></i></button>
+<iframe src="/webssh/host" frameborder="0" scrolling="no" seamless="seamless" style="display:block; width:100%; height:calc(100vh - 85px);marginx-top:40px;"></iframe>
