@@ -67,9 +67,7 @@
             <div class="collapse navbar-collapse">
                 <ul class="nav navbar-nav">
                     <li><a class="navbar-brand" style="padding-top: 17px" href="~/Env">Environment</a></li>
-                    <li><a class="navbar-brand" style="padding-top: 17px" href="~/webssh/host">SSH</a></li>
-                    <li><a class="navbar-brand" style="padding-top: 17px" href="~/DebugConsole">Bash</a></li>
-                    <li><a class="navbar-brand" style="padding-top: 17px" href="~/api/logstream" title="If no log events are being generated the page may not load.">Log stream</a></li>
+                    <li><a class="navbar-brand" style="padding-top: 17px" href="~/instances/any/webssh/host">SSH</a></li>
                 </ul>
                 @{
                     var email = httpContextAccessor.HttpContext.Request.Headers["X-MS-CLIENT-PRINCIPAL-NAME"].ToString();
