@@ -8,11 +8,11 @@ namespace Kudu.Services.Web.Tracing
 {
     public static class TraceServices
     {
-        private static readonly object _traceKey = new object();
-        private static readonly object _traceFileKey = new object();
+        private static readonly object TraceKey = new object();
+        private static readonly object TraceFileKey = new object();
 
         private static Func<ITracer> _traceFactory;
-        
+
         // CORE TODO The CurrentRequestTraceFile - Done and HttpMethod properties
         // were replaced with methods that require the caller to pass the context, as HttpContext.Current
         // no longer exists and is a bad practice anyway. CurrentRequestTracer was removed, as
@@ -20,12 +20,7 @@ namespace Kudu.Services.Web.Tracing
 
         internal static string GetRequestTraceFile(HttpContext context)
         {
-            if (context == null)
-            {
-                return null;
-            }
-
-            return context.Items[_traceFileKey] as string;
+            return context?.Items[TraceFileKey] as string;
         }
 
         internal static string GetHttpMethod(HttpContext context)
@@ -38,10 +33,7 @@ namespace Kudu.Services.Web.Tracing
             return context.Request.Method;
         }
 
-        internal static TraceLevel TraceLevel
-        {
-            get; set;
-        }
+        internal static TraceLevel TraceLevel { get; set; }
 
         public static void SetTraceFactory(Func<ITracer> traceFactory)
         {
@@ -51,37 +43,38 @@ namespace Kudu.Services.Web.Tracing
         internal static ITracer GetRequestTracer(HttpContext httpContext)
         {
             if (httpContext == null) return null;
-            return httpContext.Items[_traceKey] as ITracer;
+            return httpContext.Items[TraceKey] as ITracer;
         }
 
         internal static void RemoveRequestTracer(HttpContext httpContext)
         {
-            httpContext.Items.Remove(_traceKey);
-            httpContext.Items.Remove(_traceFileKey);
+            httpContext.Items.Remove(TraceKey);
+            httpContext.Items.Remove(TraceFileKey);
         }
 
-        internal static ITracer EnsureETWTracer(HttpContext httpContext)
+        internal static ITracer EnsureEtwTracer(HttpContext httpContext)
         {
-            var etwTracer = new ETWTracer((string)httpContext.Items[Constants.RequestIdHeader], httpContext.Request.Method);
+            var etwTracer = new ETWTracer((string) httpContext.Items[Constants.RequestIdHeader],
+                httpContext.Request.Method);
 
-            httpContext.Items[_traceKey] = etwTracer;
+            httpContext.Items[TraceKey] = etwTracer;
 
             return etwTracer;
         }
 
         internal static ITracer CreateRequestTracer(HttpContext httpContext)
         {
-            var tracer = (ITracer)httpContext.Items[_traceKey];
+            var tracer = (ITracer) httpContext.Items[TraceKey];
             if (tracer == null)
             {
-                httpContext.Items[_traceFileKey] = String.Format(Constants.TraceFileFormat, Environment.MachineName, Guid.NewGuid().ToString("D"));
+                httpContext.Items[TraceFileKey] = String.Format(Constants.TraceFileFormat, Environment.MachineName,
+                    Guid.NewGuid().ToString("D"));
 
                 tracer = _traceFactory();
-                httpContext.Items[_traceKey] = tracer;
+                httpContext.Items[TraceKey] = tracer;
             }
 
             return tracer;
         }
-        
     }
 }
