@@ -2,7 +2,6 @@
 using Kudu.Core;
 using Kudu.Services.Infrastructure;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Internal;
 using System;
 using System.IO.Abstractions;
 using System.IO.Compression;
@@ -11,6 +10,10 @@ using Microsoft.Extensions.Logging;
 using System.IO;
 using Microsoft.Extensions.DependencyInjection;
 using Kudu.Core.Infrastructure;
+using Kudu.Core.Helpers;
+using System.Threading;
+using Microsoft.AspNetCore.Http;
+using Kudu.Core.K8SE;
 
 namespace Kudu.Services.Zip
 {
@@ -18,9 +21,28 @@ namespace Kudu.Services.Zip
     // of good reusable logic in there. We could consider extracting a more basic base class from it.
     public class ZipController : VfsControllerBase
     {
-        public ZipController(ITracer tracer, IEnvironment environment)
-            : base(tracer, environment, environment.RootPath)
+        private ITracer _tracer;
+        private IEnvironment _environment;
+        public ZipController(ITracer tracer, IEnvironment environment, IHttpContextAccessor accessor)
+            : base(tracer, GetEnvironment(accessor, environment), GetEnvironment(accessor, environment).RootPath)
         {
+            _tracer = tracer;
+            _environment = GetEnvironment(accessor, environment);
+        }
+
+        private static IEnvironment GetEnvironment(IHttpContextAccessor accessor, IEnvironment environment)
+        {
+            IEnvironment _environment;
+            var context = accessor.HttpContext;
+            if (!K8SEDeploymentHelper.IsK8SEEnvironment())
+            {
+                _environment = environment;
+            }
+            else
+            {
+                _environment = (IEnvironment)context.Items["environment"];
+            }
+            return _environment;
         }
 
         protected override Task<IActionResult> CreateDirectoryGetResponse(DirectoryInfoBase info, string localFilePath)
@@ -70,7 +92,7 @@ namespace Kudu.Services.Zip
         {
             var zipArchive = new ZipArchive(Request.Body, ZipArchiveMode.Read);
             zipArchive.Extract(localFilePath);
-
+            PermissionHelper.ChmodRecursive("777", localFilePath, _tracer, TimeSpan.FromSeconds(30));
             return Task.FromResult((IActionResult)Ok());
         }
 
